+++ app/assets/javascripts/tickets.js
@@ -69,4 +69,16 @@ jQuery(function() {
     }
   });
 
+  if(jQuery('[data-lock-path]').length > 0) {
+
+    function keepLock() {
+      jQuery.ajax({
+        url: jQuery('[data-lock-path]').data('lock-path'),
+        type: 'post'
+      });
+	}
+	keepLock();
+	/* renew lock every 4 minutes */
+	setInterval(keepLock, 1000 * 60 * 4);
+  }
 });